requirements.txt
openai==1.23.2
faiss-cpu==1.7.4
numpy==1.26.4
fastapi==0.111.0
python-dotenv==1.0.1
mangum==0.17.0